pyproject.toml
[project]
name = "ai-summarizer"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "kp", email = "kou@ren-ga.jp" }
]
dependencies = [
    "selenium>=4.19.0",
    "undetected-chromedriver>=3.5.5",
    "pandas>=2.2.2",
    "easydict>=1.13",
    "filetype>=1.2.0",
    "emoji>=2.11.0",
    "mergedeep>=1.3.4",
    "pyyaml>=6.0.1",
    "talkingheads>=0.5.0",
    "python-dotenv>=1.0.1",
    "clipboard>=0.0.4",
    "slack-sdk>=3.27.1",
    "playwright>=1.43.0",
    "playwright-stealth>=1.0.6",
    "setuptools>=69.5.1",
    "kaggle>=1.6.12",
    "injector>=0.21.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.4.1",
    "mypy>=1.9.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ai_summarizer"]

[tool.mypy]
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false
disallow_untyped_calls = false

[tool.ruff]
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "W", # pycodestyle warnings
    "I", # isort
    "D", # pydocstyle
]
ignore = []
line-length = 120

extend-ignore = [
    "D105", # undocumented-magic-method
    "D107", # undocumented-public-init
    "D205", # blank-line-after-summary
    "D415", # ends-in-punctuation
    "D100", # missing-module-docstring
    "D101", # missing-module-docstring
    "D102", # missing-module-docstring
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
